refactor(analysis): split dislo.py into dislocation subpackage
- create cdiutils.analysis.dislocation subpackage
- split large dislo.py into modular modules:
  _strain_map, _clustering, _geometry, _ring, _theory, _phase_decomp, _plotting
- expose public API via dislocation.__init__
- improve maintainability and readability
